Rename model_base module to base_classes
